+++ version.json
@@ -1,6 +1,6 @@
 {
     "beta" : {
-        "r_ver" : "v3.4.4 (stable1) (beta3) (dev10)",
+        "r_ver" : "v3.4.4 (stable1) (beta3) (dev11)",
         "c_ver" : "3500105",
         "s_ver" : "3500110"
     }
