+++ views/main_css/css/main.css
@@ -120,7 +120,7 @@ a {
 }
 
 /* 위키 내용 관련 */
-#toc, .opennamuTOC {
+#toc, .opennamu_TOC {
     border: 1px solid gainsboro;
     padding: 20px;
     
@@ -146,7 +146,7 @@ pre {
     white-space: pre-wrap;
 }
 
-#toc_title, .opennamuTOCtitle {
+#toc_title, .opennamu_TOC_title {
     font-size: 1.2rem;
 }
